Intermediate state
