+++ Cargo.toml
@@ -7,18 +7,20 @@ edition = "2021"
 # atmosphere = { version = "0.1.3", features = ["sqlite"] }
 axum = { version = "0.7.4", features = ["macros", "http2"] }
 axum-extra = { version = "0.9.2", features = ["typed-header"] }
-color-eyre = "0.6.2"
 dotenv = "0.15.0"
 dotenv_codegen = "0.15.0"
-eyre = "0.6.12"
 jsonwebtoken = "9.2.0"
+miette = { version = "7.1.0", features = ["fancy"] }
 mime = "0.3.17"
 serde = { version = "1.0.196", features = ["derive"] }
 serde_json = "1.0.113"
 sqlx = { version = "0.7.3", features = ["runtime-tokio", "sqlite"] }
+thiserror = "1.0.57"
 tokio = { version = "1.36.0", features = ["full"] }
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
+utoipa = { version = "4.2.0", features = ["axum_extras"] }
+utoipa-swagger-ui = { version = "6.0.0", features = ["axum"] }
 
 [dependencies.atmosphere]
 git = "https://github.com/bmc-labs/atmosphere.git"
