+++ rustfmt.toml
@@ -1,1 +1,1 @@
-imports_granularity = "Crate"
+imports_granularity = "Module"
